requirements.txt
asgiref==3.11.1
channels==4.2.2
channels-redis==4.2.1
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
joblib==1.5.3
numpy==2.4.2
scikit-learn==1.8.0
scipy==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
gunicorn==21.2.0
psycopg2-binary
dj-database-url
uvicorn==0.35.0
whitenoise
